+++ epantry/app/src/main/java/com/example/teamfoodie/epantry/PantryUpdateActivity.java
@@ -99,7 +99,7 @@ public class PantryUpdateActivity extends AppCompatActivity implements View.OnCl
     }
 
     public void topUpIngredient() {
-        int newQuantity = ingredient.getCurrentQuantity() + ingredient.getTotalQuantity();
+        double newQuantity = ingredient.getCurrentQuantity() + ingredient.getTotalQuantity();
         ingredient.setCurrentQuantity(newQuantity);
         boolean updated = database.topUpQuantity(ingredient);
         if (updated) {
